requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-dotenv>=1.0.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
langchain-text-splitters>=0.3.0
chromadb>=0.5.0
beautifulsoup4>=4.12.0
pydantic>=2.0.0
